+++ apps/web/components/booking/types.ts
@@ -1,0 +1,17 @@
+import type { RouterInputs, RouterOutputs } from "@calcom/trpc/react";
+
+type BookingListingStatus = RouterInputs["viewer"]["bookings"]["get"]["filters"]["status"];
+
+type BookingItem = RouterOutputs["viewer"]["bookings"]["get"]["bookings"][number];
+
+export type BookingItemProps = BookingItem & {
+  listingStatus: BookingListingStatus;
+  recurringInfo: RouterOutputs["viewer"]["bookings"]["get"]["recurringInfo"][number] | undefined;
+  loggedInUser: {
+    userId: number | undefined;
+    userTimeZone: string | undefined;
+    userTimeFormat: number | null | undefined;
+    userEmail: string | undefined;
+  };
+  isToday: boolean;
+};
